Made Scripts work for multiple source code files
